Support images added to the DOM asynchronously after the page is loaded.
Rather than binding a `mouseenter` event handler to each image element individually, this commit uses a delegated event listener attached to the document.  This approach has two distinct advantages: A) it allows processing of events from elements added _after_ the handler was bound, and B) it significantly reduces overhead on pages with a large number of image elements.