Check for existence of Content-Length HTTP response header before displaying image size.
According to RFC 2616 sec. 14.13 (http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.13), the Content-Length header may not always be present in an HTTP response.  To avoid displaying `null B` to the user for images that lack this header, we should first check if the header exists.  If it doesn't, we simply omit the size in the UI.