Use unambiguous binary prefixes for image sizes (e.g., `KiB`).
The calculations for image size currently use binary math (hooray!), but the prefixes displayed in the UI are incorrect: `Kb` and `Mb` are shown instead of `KiB` and `MiB`, respectively.

The difference is subtle, but important: `Kb` and `Mb` signify `kilobit` (10^3 bits = 125 bytes) and `megabit` (10^6 bits = 125000 bytes), whereas `KiB` and `MiB` signify `kibibyte` (2^10 bytes = 1024 bytes) and `mebibyte` (2^20 bytes = 1048576 bytes).  The values in the UI are currently being displayed as the former, but are actually the latter.  Moreover, `Kb` and `KB` are ambiguous; they are often misused elsewhere, leading to user confusion.

This commit simply updates the size prefixes to unambiguous binary units.